use catch_unwind instead of threads
